+++ CHANGELOG.md
@@ -19,11 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
-## [1.1.0-beta-1] - 2026-07-15
+## [1.1.0-beta-1] - 2026-07-21
 
 ### Added
 
 - AAR050 - InfoTitleRequired: The `info.title` field must exist and not be empty.
+- AAR056 - AvroSchemaFormat: When the payload uses Avro, `schemaFormat` must be exactly `application/vnd.apache.avro;version=1.9.0`.
 
 ## [1.0.0] - 2025-12-29